requirements.txt
asgiref==3.8.1
cloudinary==1.36.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.13
django-allauth==0.57.2
django-braces==1.15.0
django-cron==0.6.0
django-crontab==0.7.1
django-model-utils==4.5.1
django-star-ratings==0.9.2
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python-crontab==3.2.0
python3-openid==3.2.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
swapper==1.3.0
urllib3==1.26.19
whitenoise==5.3.0
